% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningFunctions_Roxygen.R
\docType{package}
\name{CleaningFunctions}
\alias{CleaningFunctions}
\title{CleaningFunctions}
\description{
This package contains multiple functions to clean text

Rm_Emojis();

Rm_Hashtags();

Rm_Mentions();

Rm_Numbers();

Rm_Punctuation();

Rm_URLs();

To_Lower();

To_Singular_String()
}
\details{
To get some help concerning the functions of this package, use the following:

?Rm_Emojis

?Rm_Hashtags

?Rm_Mentions

?Rm_Numbers

?Rm_Punctuation

?Rm_URLs

?To_Lower

?To_Singular_String
}
\author{
Jean-François Savard \email{jean-françois.savard@enap.ca}
}
