% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningFunctions_R.R
\name{Rm_Punctuation}
\alias{Rm_Punctuation}
\title{Rm_Punctuation}
\usage{
Rm_Punctuation(Sstring)
}
\arguments{
\item{Sstring}{string value input}
}
\description{
this function removes punctuation
}
