% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningFunctions_R.R
\name{To_Singular_String}
\alias{To_Singular_String}
\title{To_Singular_String}
\usage{
To_Singular_String(Sstring)
}
\arguments{
\item{Sstring}{string value input}
}
\description{
This function changes all words to singular. Only applicable to a df or tibble of word tokens
}
